lib: Move power control API from hwtest
This commit moves the power control API, which was under
`test/hwtest`, to the `lib` directory. This is in preparation for
common use with the SC-Sat1 Flight software.
