+++ lib/core/Kconfig.adcs
@@ -1,0 +1,16 @@
+# Copyright (c) 2024 Space Cubics, LLC.
+# SPDX-License-Identifier: Apache-2.0
+
+menuconfig SC_LIB_CORE_ADCS
+	bool "Space Cubics Core libraries for ADCS Board"
+	default y
+	help
+	  Enable Space Cubics Core libraries for ADCS Board.
+
+if SC_LIB_CORE_ADCS
+
+module = SC_LIB_CORE_ADCS
+module-str = Space Cubics Core libraries for ADCS Board
+source "subsys/logging/Kconfig.template.log_config"
+
+endif
