+++ lib/core/Kconfig.main
@@ -1,0 +1,16 @@
+# Copyright (c) 2024 Space Cubics, LLC.
+# SPDX-License-Identifier: Apache-2.0
+
+menuconfig SC_LIB_CORE_MAIN
+	bool "Space Cubics Core libraries for MAIN Board"
+	default y
+	help
+	  Enable Space Cubics Core libraries for MAIN Board.
+
+if SC_LIB_CORE_MAIN
+
+module = SC_LIB_CORE_MAIN
+module-str = Space Cubics Core libraries for MAIN Board
+source "subsys/logging/Kconfig.template.log_config"
+
+endif
