+++ tests/hwtest/adcs/src/syshk_test.c
@@ -7,7 +7,7 @@
 #include <stdlib.h>
 #include <zephyr/kernel.h>
 #include "common.h"
-#include "pwrctrl.h"
+#include "pwrctrl_adcs.h"
 #include "temp_test.h"
 #include "cv_test.h"
 #include "imu_test.h"
