+++ tests/hwtest/main/CMakeLists.txt
@@ -10,7 +10,6 @@ project(scsat1-main)
 
 target_sources(app PRIVATE src/main.c)
 target_sources(app PRIVATE src/wdog.c)
-target_sources(app PRIVATE src/pwrctrl.c)
 target_sources(app PRIVATE src/mgnm.c )
 target_sources(app PRIVATE src/mgnm_test.c )
 target_sources(app PRIVATE src/temp.c)
